requirements.txt
flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
itsdangerous==2.2.0
click==8.1.7
werkzeug==3.0.3
python-dotenv